pyproject.toml
[tool.poetry]
name = "ml-training-template"
version = "0.1.0"
description = "A simple Machine Learning training with replication."
authors = ["ricoms"]
license = "MIT"
readme = "README.md"
packages = [{include = "ml_training_template"}]

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.6.3"
scikit-learn = "^1.2.1"
joblib = "^1.2.0"
pyyaml = "^6.0"
pillow = "^9.4.0"
pandas = "^1.5.3"
dask = {extras = ["dataframe"], version = "^2023.1.1"}


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
